float -> double; world: light->ambient

+++ projects/raytracer/BRDFs/BRDF.cpp
@@ -52,7 +52,7 @@ RGBColor BRDF::sample_f(const ShadeRec& sr, const Vector3d& wo, Vector3d& wi) co
 }
 
 
-RGBColor BRDF::sample_f(const ShadeRec& sr, const Vector3d& wo, Vector3d& wi, float& pdf) const {
+RGBColor BRDF::sample_f(const ShadeRec& sr, const Vector3d& wo, Vector3d& wi, double& pdf) const {
   return RGBColor();
 }
