+++ projects/raytracer/Cameras/Pinhole.h
@@ -12,24 +12,24 @@ class Pinhole: public Camera {
   Pinhole& operator= (const Pinhole& rhs);	
   virtual ~Pinhole();
   virtual Pinhole* clone() const;  
-  void set_view_distance(const float d);
-  void set_zoom(const float zoom);
+  void set_view_distance(const double d);
+  void set_zoom(const double zoom);
   Vector3d get_direction(const Vector2d& p) const;
   virtual void render_scene(const World* w, const char* image_file);
 		
  private:
-  float	d;                      // view plane distance
-  float	zoom;                   // zoom factor
+  double	d;                      // view plane distance
+  double	zoom;                   // zoom factor
 
 };
 
 
-inline void Pinhole::set_view_distance(const float d) {
+inline void Pinhole::set_view_distance(const double d) {
   this->d = d;
 }	
 
 
-inline void Pinhole::set_zoom(const float zoom) {
+inline void Pinhole::set_zoom(const double zoom) {
   this->zoom = zoom;
 }
